% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmatch.R
\name{assess_match_quality}
\alias{assess_match_quality}
\title{Assess Matching Quality (Academic Standard)}
\usage{
assess_match_quality(
  match_res,
  stack_candidates,
  dt,
  match_vars,
  id_col,
  time_col
)
}
\description{
Assess Matching Quality (Academic Standard)
}
