% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmatch.R
\name{expand_matched_stacks}
\alias{expand_matched_stacks}
\title{Expand Matched Cross-Section to Full Stacks}
\usage{
expand_matched_stacks(
  dt,
  matched_data,
  id_col,
  time_col,
  k_pre,
  k_post,
  match_by = -1
)
}
\description{
Expand Matched Cross-Section to Full Stacks
}
