% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmatch.R
\name{label_candidates}
\alias{label_candidates}
\title{Label Treatment and Control Stack Candidates}
\usage{
label_candidates(
  dt,
  id_col,
  time_col,
  event_col,
  k_pre,
  k_post,
  k_freeze,
  match_by = -1
)
}
\description{
Identifies valid stack candidates by enforcing strict data continuity and
clean history/future constraints.
}
