% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmatch.R
\name{match_candidates}
\alias{match_candidates}
\title{Cross-Sectional Matching for Stacked DiD}
\usage{
match_candidates(dt, stack_candidates, id_col, time_col, match_vars, ...)
}
\description{
Performs exact cohort matching + 1-to-K matching using MatchIt.
}
