% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmatch.R
\name{plot_event_dynamics}
\alias{plot_event_dynamics}
\title{Plot Event Dynamics (Lasagna Plot with Overlay)}
\usage{
plot_event_dynamics(
  dt,
  dt_matched = NULL,
  id_col,
  time_col,
  event_col,
  n_bins = 4L,
  sort_by_intensity = FALSE
)
}
\description{
Plot Event Dynamics (Lasagna Plot with Overlay)
}
